friends tab updated
