+++ lib/features/friends/AddFriends/QRCode.dart
